refactor: initialized refactor of modules
