+++ frontend/src/hir/modules/symbols.rs
